python-starter/requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
motor
python-dotenv
pytest-asyncio
bson
pydantic_mongo
